+++ src/RcsbFv/RcsbFvTooltip/RcsbFvTooltipManager.ts
@@ -75,7 +75,7 @@ export class RcsbFvTooltipManager {
     showTooltipDescription(d: RcsbFvTrackDataElementInterface){
         this.tooltipDescriptionDiv.textContent = "";
         this.tooltipDescriptionDiv.style.visibility = "hidden";
-        const description = this.tooltip.showTooltipDescription(d);
+        const description = this.tooltip.showTooltipDescription?.(d);
         if(!description)
             return;
         this.tooltipDescriptionDiv.append(description);